requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.11.2
jmespath==1.0.1
Pillow==9.3.0
pip==23.0.1
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
setuptools==65.5.1
six==1.16.0
sqlparse==0.4.3
psycopg2-binary==2.9.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.14
wheel==0.37.1
whitenoise==6.3.0
gunicorn==20.1.0
